requirements.txt
asyncio
crewai==1.3.0
autogenstudio==0.4.2.2
autogen-agentchat==0.5.7
autogen-core==0.5.7
autogen-ext==0.5.7
alembic==1.17.1
anthropic==0.72.0
arxiv==2.3.0
beautifulsoup4==4.14.2
chromadb==1.3.4
cohere==5.20.0
docx2txt==0.9
faiss-cpu==1.12.0
fastapi==0.121.0
flask==3.1.2
gradio==5.49.1
groq==0.33.0
datasets
datasets[audio]
datasets[vision]
ipykernel==7.1.0
keras==3.12.0
langchain==1.0.3
langchain-anthropic==1.0.1
langchain-chroma==1.0.0
langchain-community==0.4.1
langchain-core==1.0.3
langchain-cohere==0.5.0
langchain-google-genai==3.0.1
langchain-groq==1.0.0
langchain-huggingface==1.0.1
langchain-openai==1.0.2
langchain-pinecone==0.2.13
langchain-tavily==0.2.13
langchain-text-splitters==1.0.0
langfuse==3.9.0
langgraph==1.0.2
langgraph-cli==0.4.7
langgraph-cli[inmem]
langgraph-prebuilt==1.0.2
langgraph-runtime-inmem==0.12.0
langgraph-supervisor==0.0.30
langserve==0.3.3
langsmith==0.4.41
matplotlib==3.10.7
numpy==2.2.1
openai==2.7.1
passlib[bcrypt]==1.7.4
bcrypt==5.0.0
pdfminer.six==20250506
pdfplumber==0.11.7
pinecone==7.3.0
pinecone-client==6.0.0
pydantic[email]==2.11.1
pydantic-settings==2.11.0
pypdf==6.1.3
pymupdf==1.26.6
python-dotenv==1.2.1
python-jose[cryptography]==3.5.0
multipart==1.3.0
python-multipart
psycopg2-binary==2.9.11
pyJWT
rank-bm25
requests==2.32.5
scikit-learn==1.7.2
seaborn
sentence-transformers==5.1.2
sqlalchemy==2.0.44
streamlit==1.51.0
tavily-python==0.7.12
tenacity==9.1.2
tensorflow==2.20.0
torch==2.9.0
tqdm==4.67.1
transformers==4.57.1
unstructured==0.18.15
uvicorn==0.38.0
wikipedia==1.4.0
nltk==3.9.2
opencv-python==4.12.0.88